Fixed a bug happening in build causing the game to not start properly

+++ Assets/Scripts/DataPersistence/StartButton.cs
@@ -1,14 +1,18 @@
+using Microsoft.Cci;
 using System.Collections;
 using System.Collections.Generic;
 using System.IO;
 using UnityEngine;
+using UnityEngine.SceneManagement;
 
 public class StartButton : MonoBehaviour
 {
     public void OnStart()
     {
-        string[] filePaths = Directory.GetFiles(Application.persistentDataPath); 
-        
-        foreach (string filePath in filePaths) File.Delete(filePath);
+        if (Directory.Exists(Path.Combine(Application.persistentDataPath, "Data")))
+        {
+            string[] filePaths = Directory.GetFiles(string.Concat(Application.persistentDataPath, "/Data"));
+            foreach (string filePath in filePaths) File.Delete(filePath);
+        }
     }
 }
